Slightly fixed some unbeaty inside WAL-logs implementation.
We should use only 1 syscall to write log into Disk and then we can
memcpy data into right structure.

Fixed all leaks in wal-log tests. (Just add destroy-* instructions)

Reworked compilation of message.c. As this lib is required for
operations_queue and WAL-logs, it is standalone static library. It
still needs compilation of whole server sources but it can be fixed
moving message.h and message.c into own subdirectory

+++ src/wal/wal.h
@@ -36,8 +36,8 @@ typedef struct {
 } wal_logger_t;
 
 typedef struct {
-	uint64_t size;
-	char*    ptr;
+	size_t size;
+	char*  ptr;
 } binary_record_t;
 
 typedef struct {
